services/aggregator-service/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-multipart
PyJWT
pydantic>=2.0.0
